Refactoring changes after rebase
